Import 2.3.48pre2

+++ arch/mips/sgi/kernel/indy_hpc.c
@@ -1,4 +1,4 @@
-/* $Id: indy_hpc.c,v 1.6 1999/05/07 22:34:31 ulfc Exp $
+/* $Id: indy_hpc.c,v 1.9 1999/12/04 03:59:00 ralf Exp $
  *
  * indy_hpc.c: Routines for generic manipulation of the HPC controllers.
  *
@@ -8,8 +8,8 @@
 #include <linux/init.h>
 
 #include <asm/addrspace.h>
-#include <asm/sgihpc.h>
-#include <asm/sgint23.h>
+#include <asm/sgi/sgihpc.h>
+#include <asm/sgi/sgint23.h>
 #include <asm/sgialib.h>
 
 /* #define DEBUG_SGIHPC */
@@ -18,7 +18,7 @@ struct hpc3_regs *hpc3c0, *hpc3c1;
 struct hpc3_miscregs *hpc3mregs;
 
 /* We need software copies of these because they are write only. */
-unsigned long sgi_hpc_write1, sgi_hpc_write2;
+unsigned int sgi_hpc_write1, sgi_hpc_write2;
 
 /* Machine specific identifier knobs. */
 int sgi_has_ioc2 = 0;
