+++ include/asm-arm/arch-sa1100/io.h
@@ -20,7 +20,7 @@
  * Generic virtual read/write
  */
 #define __arch_getb(a)		(*(volatile unsigned char *)(a))
-#define __arch_getl(a)		(*(volatile unsigned long *)(a))
+#define __arch_getl(a)		(*(volatile unsigned int  *)(a))
 
 extern __inline__ unsigned int __arch_getw(unsigned long a)
 {
@@ -33,7 +33,7 @@ extern __inline__ unsigned int __arch_getw(unsigned long a)
 
 
 #define __arch_putb(v,a)	(*(volatile unsigned char *)(a) = (v))
-#define __arch_putl(v,a)	(*(volatile unsigned long *)(a) = (v))
+#define __arch_putl(v,a)	(*(volatile unsigned int  *)(a) = (v))
 
 extern __inline__ void __arch_putw(unsigned int value, unsigned long a)
 {
